moved all significant logic out of game_state
