+++ aimmo-game/tests/test_simulation/test_simulation_runner.py
@@ -8,9 +8,9 @@ from simulation.game_state import GameState
 from simulation.location import Location
 from simulation.simulation_runner import ConcurrentSimulationRunner
 
-from .dummy_avatar import (DeadDummy, DummyAvatarManager, MoveEastDummy,
-                           MoveNorthDummy, MoveSouthDummy, MoveWestDummy,
-                           WaitDummy)
+from .dummy_avatar import (DeadDummy, DummyAvatar, DummyAvatarManager,
+                           MoveEastDummy, MoveNorthDummy, MoveSouthDummy,
+                           MoveWestDummy, WaitDummy)
 from .maps import InfiniteMap
 from .mock_communicator import MockCommunicator
 
@@ -77,6 +77,27 @@ class TestSimulationRunner(unittest.TestCase):
         avatar = self.game_state.avatar_manager.avatars_by_id[7]
         self.assertEqual(avatar.location.x, 2)
         self.assertEqual(avatar.location.y, 2)
+
+    def test_remove_avatar(self):
+        self.construct_simulation_runner(
+            [DummyAvatar, DummyAvatar],
+            [Location(0, 0), Location(1, 1)]
+        )
+
+        avatar = self.get_avatar(1)
+        avatar.marked = True
+
+        self.simulation_runner.remove_avatar(0)
+
+        self.assertNotIn(0, self.avatar_manager.avatars_by_id)
+        self.assertEqual(self.game_state.world_map.get_cell(Location(0, 0)).avatar, None)
+
+        self.assertTrue(self.avatar_manager.avatars_by_id[1].marked)
+        self.assertTrue(self.game_state.world_map.get_cell(Location(1, 1)).avatar.marked)
+
+    def test_remove_non_existant_avatar(self):
+        self.construct_simulation_runner([],[])
+        self.simulation_runner.remove_avatar(10)
 
     def test_updates_map_with_correct_num_avatars(self):
         self.construct_simulation_runner([],[])
